@@ -1,7 +1,7 @@
-export const githubOwner = 'peerjako-aws';
-export const repoName = 'cdk-ecs-cicd';
-export const gitDevBranch = 'dev'
-export const gitProdBranch = 'master'
-export const awsSecretsGitHubTokenName = 'github-personal-access-token';
-export const ssmImageTagParamName = 'penguinapp-imagetag';
-export const stagingValidationEmail = 'validation@test.com';
+export const githubOwner = "sonuabraham-aremedia";
+export const repoName = "cdk-ecs-cicd";
+export const gitDevBranch = "dev";
+export const gitProdBranch = "master";
+export const awsSecretsGitHubTokenName = "githubaccesstoken";
+export const ssmImageTagParamName = "penguinapp-imagetag";
+export const stagingValidationEmail = "sonu.abraham@aremedia.com.au";
