@@ -1,6 +1,12 @@
-import { IRepository } from '@aws-cdk/aws-ecr';
-import { CfnTaskDefinition, ContainerDefinition, ContainerImage, ContainerImageConfig } from '@aws-cdk/aws-ecs';
-import { CfnParameter, Lazy } from '@aws-cdk/core';
+import { IRepository } from "aws-cdk-lib/aws-ecr";
+import {
+  CfnTaskDefinition,
+  ContainerDefinition,
+  ContainerImage,
+  ContainerImageConfig,
+} from "aws-cdk-lib/aws-ecs";
+import { CfnParameter, Lazy } from "aws-cdk-lib/core";
+import cdk = require("aws-cdk-lib/core");
 
 export class PipelineContainerImage extends ContainerImage {
   public readonly imageName: string;
@@ -9,14 +15,18 @@ export class PipelineContainerImage extends ContainerImage {
 
   constructor(repository: IRepository) {
     super();
-    this.imageName = repository.repositoryUriForTag(Lazy.stringValue({ produce: () => this.parameter!.valueAsString }));
+    this.imageName = repository.repositoryUriForTag(
+      Lazy.string({ produce: () => this.parameter!.valueAsString })
+    );
     this.repository = repository;
   }
 
   public bind(containerDefinition: ContainerDefinition): ContainerImageConfig {
-    this.repository.grantPull(containerDefinition.taskDefinition.obtainExecutionRole());
-    this.parameter = new CfnParameter(containerDefinition, 'PipelineParam', {
-      type: 'String',
+    this.repository.grantPull(
+      containerDefinition.taskDefinition.obtainExecutionRole()
+    );
+    this.parameter = new CfnParameter(containerDefinition, "PipelineParam", {
+      type: "String",
     });
     return {
       imageName: this.imageName,
@@ -25,10 +35,12 @@ export class PipelineContainerImage extends ContainerImage {
 
   public get paramName(): string {
     // return cdk.Token.asString(this.parameter!.logicalId).toString();
-    return Lazy.stringValue({ produce: () => this.parameter!.logicalId });
+    return Lazy.string({ produce: () => this.parameter!.logicalId });
   }
 
-  public toRepositoryCredentialsJson(): CfnTaskDefinition.RepositoryCredentialsProperty | undefined {
+  public toRepositoryCredentialsJson():
+    | CfnTaskDefinition.RepositoryCredentialsProperty
+    | undefined {
     return undefined;
   }
 }
